Add mention of releasenotes to the PTI
We expect this of OpenStack projects now. Be explicit about it.

Use the same overall structure as is used for Sphinx. Because we expect
reno requirements to be a much more consistent set, make listing them
explicitly in the repo optional. A reno releasenotes job should be able
to know what it needs to install to generate reno docs without any
specific instructions from the repo itself.

Change-Id: I1cfa2216451d056c1bdedd67186131c5e0901ecb