OpenStack-Helm - Helm charts for OpenStack
The mission of OpenStack-Helm[1] is to provide a collection of Helm charts that
simply, resiliently, and flexibly deploy OpenStack and related services
on Kubernetes.  It allows the Helm client to be used directly to deploy
OpenStack components, and as such is comarable to OpenStack-Ansible
(providing playbooks for Ansible) or Puppet OpenStack (providing modules
for Puppet).  The project leverages native Kubernetes constructs and allows
OpenStack projects to be installed in combinations and configurations per
unique operator use cases.

OpenStack-Helm has a goal to facilitate pushbutton automated deployments, and
rolling upgrades with no impact to tenant workloads.  There is a one-to-one
correspondence between OpenStack projects and Helm charts, and charts are fully
independent of one another.  This allows for use cases from traditional
OpenStack clouds, to deployment of standalone components, and combinations
in between.

The project provides reasonable, usable configuration by default, and allows
for deployment-time overrides of individual values or a chart's entire values
config.  Dependent service endpoints (e.g. the Keystone API) default to
Kubernetes cluster-internal DNS names, but can be overridden to point at
external instances.  Multiple, alternate SDN and SDS solutions will be
integrated, and can be configured at deploy-time.  The choice of container
images is configurable as well, and can be mixed and matched on a
chart-by-chart basis per operator needs.  The project aims to be
production-grade out-of-box, including configuration for highly available
services, resilient storage for the control plane (e.g. Ceph), and security
features such as cluster-internal TLS.

OpenStack-Helm will follow the OpenStack independent release model[2], and will
ensure Helm charts integrate with an OpenStack release once it is released.
The team plans to change to cycle-trailing releases when ready, but will
need to close the gap between Newton (current-state) and master first.  In the
meantime, it will model the cycle-trailing branch strategy as it catches up.
The team is interested in feedback on this plan and advice from the TC.

A number of OpenStack deployment projects pave the way for OpenStack-Helm,
including OpenStack Ansible[3], Kolla-Kubernetes[4], OpenStack-Chef[5],
and Puppet OpenStack [6].  OpenStack-Helm occupies a unique space in its
exclusive use of Kubernetes-native constructs during the deployment process,
including Helm packaging and gotpl templating.  This tradeoff enables it to be
as lightweight and simple as possible, at the expense of operator choice of
deployment toolsets.  The project team is eager to work with other OpenStack
deployment projects to solve common needs and share learnings.

The OpenStack-Helm project and the team follow the Four Opens:

* Open Source: The project source code[7], specifications[8] and
documentation[9] are open source and are developed in the community under the
Apache 2.0 license.

* Open Design: Design decisions are discussed and formalized via
specifications, as well as openly discussed in the IRC chat room[10] and the
team meetings[11].  The team had informal meetings at the Atlanta and
Denver PTGs, as well as a presentation[12], hands-on workshop, and a related
cross-project Forum at the Boston Summit.  More of each are planned for Sydney.

* Open Development: All OpenStack-Helm code is reviewed and hosted in
OpenStack Infra[7]. blueprints and bugs are tracked by Launchpad[13].
Quality gates are run to ensure OpenStack deploys and runs with every
patch set, and integration of standard Rally and Tempest tests is underway.

* Open Community:  The project team and core reviewer team[14] represent
multiple employers, and are eager to groom new team members.  Project
discussions and Q&A are frequently held in the public IRC chat room, and
all are welcome.

1. https://wiki.openstack.org/wiki/Openstack-helm
2. https://releases.openstack.org/reference/release_models.html
3. https://wiki.openstack.org/wiki/OpenStackAnsible
4. https://wiki.openstack.org/wiki/Kolla
5. https://wiki.openstack.org/wiki/Chef
6. https://docs.openstack.org/puppet-openstack-guide/latest/
7. https://git.openstack.org/cgit/openstack/openstack-helm/
8. https://git.openstack.org/cgit/openstack/openstack-helm/tree/doc/source/specs
9. http://openstack-helm.readthedocs.io/
10. http://eavesdrop.openstack.org/irclogs/%23openstack-helm/
11. http://eavesdrop.openstack.org/irclogs/%23openstack-meeting-5/%23openstack-meeting-5.2017-10-03.log.html
12. https://www.openstack.org/videos/search?search=openstack-helm
13. https://launchpad.net/openstack-helm
14. https://review.openstack.org/#/admin/groups/1749,members

Change-Id: I70995640b6b760e7bfd7f1ec407b8004c855203f