LOCI - OCI images for OpenStack
The mission of LOCI is to provide tooling that can output OCI[0]
compliant images in a CI/CD friendly way. A major goal, which we have
achieved, is to be able to build OpenStack images in a completely
air-gapped environment. Currently, the LOCI gates work completely by
consuming OpenStack-Infra mirrors for packages, pypi, and DockerHub
images.

LOCI has always been close to the OpenStack ecosystem. We are one of the
first ones to use secrets in zuulv3 outside of Infra themselves. We also
have worked with other projects to produce images per commit to the
services[1][2]. We extended openstack-infra/bindep[3] to support a new
syntax that allowed us to further simplify the LOCI build process while
making auditing of packages easier[4].

Some of the major benefits of LOCI images include:
  * Standard tooling for building (currently using Docker build)
  * Fast build times, most services output an image under 2 minutes
  * Small size, the LOCI Keystone image is under 40MB on Debian
  * CI/CD friendly, LOCI builds services entirely seperate from each
    other, so you can build Keystone seperately from Neutron without
    losing any benefits listed above
  * python3 support (though not all OpenStack services support python3)

LOCI itself only provides the tooling to build the images. It has no
deployment engine, and never will. The consumers of LOCI include several
offical OpenStack projects, namely Cinder[5] and OpenStack-Helm[6].
Outside of this, there are at least two companies using LOCI images
internally. Additionally, individuals such as myself and Chris Hoge
(hogepodge) use these images in personal deployments.

We will be implementing two periodic jobs that will reconsume zuulv3
jobs from Kolla-Ansible and OpenStack-Helm when those gates are ready.
This will allow us to demonstrate the functionality of LOCI images in
two seperate OpenStack projects on a regular basis. We have discussed
this with both PTLs from both communities and both are happy to work
with this us to make this happen.

While LOCI has been around for a year, we will be making our debut 1.0.0
stable tag for the OpenStack Queens release.

We are working on moving our meeting to one of the #openstack-meeting
channels to be in compliance with the OpenStack Project Team guidelines
as mentioned here[7][8].

To address the "4 opens":
  * Open Source: check! All code is license compatible
  * Open Commnuity:
      - We have not held any elections yet, but given acceptance by the
        TC we would do so at the next available cycle.
      - We are not, as a group, affiliated with any individual company.
        In fact, the majority of the work done in LOCI was on personal
        time for the contirbutors
  * Open Development:
      - We have been using the OpenStack eco system extensively
      - We are cooperating with existing projects as shown by the points
        in previous paragraphs as well as the links below
  * Open Design:
      - We have, in an unofficial capacity, done most of our design work
        at the PTGs or Summits. We would love to do this in an official
        capaciity!
      - We primarily use IRC for collaboration, but I do monitor the
        mailing list. I suspect as time moves forward, and development
        slows, the mailing list will get used more, as has been the case
        with other projects.

[0] https://www.opencontainers.org/
[1] https://review.openstack.org/#/c/512398/
[2] https://review.openstack.org/#/c/512793/
[3] https://review.openstack.org/#/c/506502/
[4] https://review.openstack.org/#/c/512063/3/bindep.txt
[5] https://github.com/openstack/cinder/tree/master/contrib/block-box
[6] https://review.openstack.org/#/c/472961/
[7] https://docs.openstack.org/project-team-guide/open-community.html#public-meetings-on-irc
[8] https://review.openstack.org/#/c/512471/

Change-Id: I4152237a2fbde5f827aae4524f87e7bf2798e956