split up sahara deliverables
The additional release job validation in the releases repo depends on
all of the components of a deliverable being similar enough that they
use the same release jobs.

The sahara-dashboard component uses a special job variant that the
other components do not use. The sahara-extra and
sahara-image-elements components are not services either.

This patch divides up the deliverables and copies all of the relevant
tags to the new items for consistency.

Change-Id: I461ca8acf34b545cc704a5a36ace061079959227
Signed-off-by: Doug Hellmann <doug@doughellmann.com>