Add supports-accessible-upgrade tag to ironic*
This adds the "supports-accessible-upgrade" tag to ironic and
ironic-inspector. They both meet the criteria for this tag:

https://governance.openstack.org/tc/reference/tags/assert_supports-accessible-upgrade.html#requirements

During an ironic upgrade, the bare metal nodes that end users access
are unaffected. this is tested in gate, via the grenade jobs. For
example, in the 'ironic-grenade-dsvm' job, nova is used to get a bare
metal instance from old ironic. After old ironic is shut down and
before new ironic is started up, grenade runs the resource phase:
'verify_noapi pre-upgrade'. Although ironic's verify_noapi() is a
noop [1], nova comes to the rescue with it's verify_noapi() [2]
where it pings its bare metal instance (which it got from ironic).

ironic-inspector only deals with bare metal nodes prior to them being
available to end users, so it doesn't touch any resource that is
accessible to the end user.

[1] https://github.com/openstack/ironic/blob/7f924cf30f546ae76eb149ae86201073ded396b2/devstack/upgrade/resources.sh#L123

[2] https://github.com/openstack-dev/grenade/blob/dbee4eed16cc0b2a2143f18936d6c85fd3940da0/projects/60_nova/resources.sh#L136

Change-Id: I20941939e9f2331c6a0983e3c14c988cae936bd2