Remove stable:follows-policy from Kolla
During the OpenStack Summit in Sydney, we discussed about what operators
expect from the stable policy.
One of the outputs, documented here [1], was that projects like
installers don't need the tag.
The tag has been created for service projects with an API.

This patch removes the tag from Kolla, which was the only remaining
deployment tool project with this tag.
We'll also have to update the Software Catalog (on openstack.org) to
make sure it doesn't show the tag as missing.

[1] https://etherpad.openstack.org/p/SYD-stable-policy

Change-Id: If4803d14db0c51c83789df302fdb0471d72fb824