Stable policy applies to main OpenStack components
The stable:follows-policy tag is meant to provide information
about components of an OpenStack cloud, but does not apply so
well to downstream packaging or lifecycle management tools,
where the expectations are different[1].

This proposal recenters the tag on those "cloud software"
components, and removes the expectation that lifecycle management
tools or SDKs are expected to get that tag to show maturity.

If this is approved, the "Software Navigator" on openstack.org
will be updated so that missing that tag no longer adversely
affects the maturity score of deployment tools.

[1] http://lists.openstack.org/pipermail/openstack-dev/2017-October/123624.html

Change-Id: I1c1ad6b4843a04b7b5debb092fa49bffad4b12ac