Remove releasenotes/requirements.txt
In change Iaa917d4f126d24c701cd3b8b0edfba4d4e0ca9c1 we have the
proposal to not to install releasenotes/requirements.txt as projects
should not really need any extra requirements to produce releasenotes
other than reno itself being installed and available.

Remove the references to this requirements.txt file so the CI
implementation does not get out of sync with policy.  As noted by
AJaeger in the other change, not repo uses this as yet.

Change-Id: Ieab3dfde7854d76dded6b1f8dd3aab1f1071b208
Needed-by: Iaa917d4f126d24c701cd3b8b0edfba4d4e0ca9c1