Rename shade team to OpenStackSDK
Align the name to be similar to OpenStackClient and to be clear what the
team does.

This doesn't indicate Python, even though the mission is currently
limited to python. I don't believe there is a large appetite to move
non-python OpenStack SDK work into OpenStack directly, but at such a
time as such interest is shown, we could either rename the team again
(fairly low-cost) or consider whether being a combined effort makes
sense. In any case, this best describes the state of the world today.

Change-Id: Ic7415e25c4f04f49f0feb538348254355830c846