Add an openstack/self-healing-sig repository
The Self-healing SIG is interested in using Storyboard for task
tracking purposes. At the moment our means of avoiding name
collisions is to make sure each project in SB has a matching
repository in Gerrit, so create the necessary repo as a placeholder
(it may be used for revision control of documents or other artifacts
pertaining to the SIG in the future, but they have no actual plans
for it at this time).

https://wiki.openstack.org/wiki/Self-healing_SIG
http://lists.openstack.org/pipermail/openstack-sigs/2017-December/000198.html

Change-Id: Ie2bba5e8914eb5422974b968183498471fae37cb
Depends-On: If4fed236b6a679a635e17984f1cd07e98e21758c