Allow public polling to choose release names
Release naming polls originally were conducted as public polls
in Launchpad (requiring only a Launchpad account). Then we moved
to CIVS-driven Condorcet polls, with personal voting links to
all Foundation members (requiring a Foundation membership to vote).

Due to limitations in the CIVS service, it's currently not practical
to send Condorcet voting links to every Foundation member. Instead
we'll use the public voting feature of CIVS, and allow anyone with
the link to cast their vote.

That opens the vote to ballot stuffing, but the choice is not that
critical, the proposed names all follow strict criteria, and there
are many safeguards in place to exclude bad names (from a legal or
marketing standpoint). So it feels safe enough to use public voting,
originally disseminating the voting URL through the openstack
mailing-lists.

Change-Id: I2f4b7fb24ccfb54403d0dedc0a757af3c6769d61