Remove editor mode setting from goals template
Not everyone uses the same text editors, and it's annoying to have
one person's preferred editor configuration encoded into files meant
for others to edit. Remove the emacs mode variable comment from the
top of the goals template, since not only was it setting someone's
editor preference on a file but doing so on a file meant as an
example for others to copy. If people want to set particular editor
modes automatically, they can do so based on the filename pattern
(we're pretty consistent about setting clear suffixes).

Change-Id: I60932b0a2d86f23de6d80637b04f3c0f76830b70