Add resolution about CI for external projects
This resolution is to establish agreement that we think extending
Zuul to run integration tests and report on non-OpenStack projects
(e.g., Ansible, kubernetes, etc.) is in the interest of the community,
and what the scope of such an effort should be.  I plan to follow
it up with specific instructions in the infra-manual describing
the process to set up such testing.

Change-Id: Ia7dc06cade7e436a4df64f4f0d55327ea6c78b81