Update status of 'API deployed in uwsgi' goal
The Pike release of Glance contained a release note indicating
the Glance should not be deployed under uwsgi in production.  That
same advice is being given for Queens.  Add a note to the goal
completion section for that goal so that people who don't read the
release notes are aware of this.

Change-Id: Iab7310751874392c8cf787b6cd748aa55a7b7477