Update new projects reference WRT IRC meetings.
In I631e2597bf818c5e4b2f813da14764d2cb55d0ce the TC passed
resolutions/20170718-allow-scheduling-meetings-on-team-channels which
allows teams to host meetings in per-project IRC channels.

However the new-projects-requirements still stated that meeting needed
to be held in one of the central meeting channels.  This change updates
that document in light of the resolution.

Change-Id: Ib8febcc6b80bc36833f2040b1cc1c5eb15841655