Add systemd roles to OpenStack-Ansible
The systemd_mount and systemd_service roles are a way for
OpenStack-Ansible to not repeat ourselves in the code.

It can also be used by other members of the openstack community,
and is therefore not mapped to our main deliverable.
We are still curating it in our group: these roles will be important
for deploying a cloud with openstack-ansible.

Depends-On: https://review.openstack.org/#/c/552861
Change-Id: I098ccc03ff2e4a4bcb65e1f376f2c712b10c9c25