Add python_venv_build ansible role to OSA deliverables
In OpenStack-Ansible python venvs are built, packaged
and consumed at run-time. Currently this uses the same
code across multiple roles. This role will be used by
OpenStack-Ansible to replace the duplicated code in
every OSA role but could also be used by any other
project who may wish to.

The role is registered as an independent deliverable as
it is not expected to require frequent releasing.

Depends-On: https://review.openstack.org/553392
Change-Id: Ic4bad4c0cebe018c4a27e0cc79faf9186b7707de