Adjutant official project status
Adjuant is a service built to help manage certain elements of operations
processes by providing micro APIs around complex underlying workflow. It
started life as a system to manage sign ups to a public cloud, and grew
into a more generic and flexible system into which deployers can add
useful APIs for their users to consume.

The project history can be found on the docs[1], as well as some guide
lines which explain what the scope of the project is, and how we manage
how vague it is[2]. Then there is a section going over what is built
into Adjutant[3].

The project was built to be integrated from the beginning with
OpenStack, using KeystoneMiddleware for auth, will be moving to oslo
policy in the near future, and we hope to switch to using the
OpenStackSDK for all OpenStack API interactions.

Adjutant meets all of the new project requirements easily. The source
is Apache License 2.0, and all libraries used are opensource. While the
project was started internally at Catalyst we use Launchpad for bug and
blueprints tracking and all of us are developers already working
upstream with OpenStack. Code review and gated testing is done on the
OpenStack infrastructure, and the goal is always to support features in
other projects first (see guide-lines doc[2]). We want to build Adjutant
to be useful to the community and have taken care to build ways to keep
company specific logic out of the service by providing plugin mechanisms.

Catalyst Cloud[4] is a company dedicated to opensource, and we always
prefer working with others to achieve a goal that works for everyone
rather than writing internal only solutions. Almost all our work is
opensource where applicable, and we are active in the community.

Catalyst Cloud runs Adjutant in production (very near to master), and we
know of at least another company using it for password resets, as well
as interest from a few more. For Catalyst Cloud it handles our full
automated sign up process, and we run almost all of the default Tasks as
defined in the core codebase for user management.

[1]: https://adjutant.readthedocs.io/en/latest/history.html
[2]: https://adjutant.readthedocs.io/en/latest/guide-lines.html
[3]: https://adjutant.readthedocs.io/en/latest/features.html
[4]: https://catalystcloud.nz/

Change-Id: I0d119fa26b7ed8969870ad0c3f405e0ac3df98e3