Resolution for minimal SIG governance
We have been operating SIGs for some time now with a very loose
governance model. We took this approach on purpose to give folks
space and opportunity to decide what they believed constituted
a SIG in hopes of learning how we can apply SIGs to our community
needs.

However at this point there is some confusion and concern regarding
SIG governance. Since SIGs explicitly do not sit below any existing
governance structure, there is no default solution to fall back on.
We need to define a clear way to address issues before any conflict
arises. As was the case elsewhere, we expect the mere existence of
a clear decision-making structure to prevent most conflicts from
happening.

This resolution proposes a minimal, simple governance model to
address such cases, introducing a SIGs admin group. The group
is initially formed of the Meta SIG co-leads: Melvin Hillsman
as UC representative and Thierry Carrez as TC representative.
The committees are of course free to choose new members whenever
they see fit.

A similar resolution was proposed to the User Committee, which
approved it:

http://eavesdrop.openstack.org/meetings/uc/2018/uc.2018-03-12-14.05.log.html#l-74

Change-Id: Ibc2336980739db38f64a42b8ba2adeacff000a13
Co-Authored-By: Melvin Hillsman <mrhillsman@gmail.com>