Add puppet-senlin to Puppet OpenStack
"Senlin is a clustering service for OpenStack clouds. It creates and operates
clusters of homogeneous objects exposed by other OpenStack services. The
goal is to make the orchestration of collections of similar objects easier.

Senlin provides RESTful APIs to users so that they can associate various
policies to a cluster.  Sample policies include placement policy, load
balancing policy, health policy, scaling policy, update policy and so on.

Senlin is designed to be capable of managing different types of objects. An
object's lifecycle is managed using profile type implementations, which are
themselves plugins."

More information about Senlin project:
https://wiki.openstack.org/wiki/Senlin

We started to work on the Puppet senlin module that installs and
configures Senlin services.

Change-Id: I53b62c6874da44215a89b56f57c44f954aeab057
Depends-On: Ic4272d28016b95006f6c31c41f60c9111ebb2c91