Adjust TC election target to 6 weeks before summit
The existing 3 weeks prior to summit target for TC elections can be
problematic for travel planning for candidates who might only go to
summit if they win their election, or might plan a different length
of trip depending on their role(s) in the community. This change
makes the target for the election to be six weeks prior to summit
to ease that planning.

In addition to helping with travel concerns, it also means that any
newly elected TC members will be more involved in planning for the
forum at the summit.

If approved this change would go in effect for the second election
in 2018. The first election of 2018 is already scheduled.

Change-Id: Ie1a7d9bae28864fb7413d8df2daa67abc5c70778