provide more detail about the expectations we place on goal champions
At least one of our goal champions was under the mistaken impression
that they were responsible for writing all of the patches for teams
that did not complete the goal themselves. This patch adds more detail
about what responsibility a champion has, and what things are not
their responsibility.

Story: #2001706
Task: #8699
Change-Id: Ie00f793378ed2ec870210ffd480f29a510a427d5
Signed-off-by: Doug Hellmann <doug@doughellmann.com>