Doug Hellmann nomination for Chair
I would be honored to serve as chair for the TC for this session.

With Thierry stepping down as chair, this session will be the first
time the TC as it stands today is chaired by someone from outside of
the Foundation staff. This is an important step for us, to show not
only that our principle of Changes in Leadership are Good applies to
the TC, but that our leaders can and will place the needs of the
community first. Over the 6 years I have been contributing to
OpenStack and the 4+ years I have served on the TC, I have been
employed by 3 separate companies. I hope my history of service within
the community reassures everyone of my continued impartiality.

The nature of the TC and the issues it has had to address have changed
in significant ways several times while I have been involved.

When the TC addressed fundamental questions about what types of
projects should be included in the community, and OpenStack expanded
from a simple IaaS project to our current mission of producing the
ubiquitous Open Source Cloud Computing platform, I helped navigate
that change, first as a member of the Ceilometer team (one of the
first teams to be added to OpenStack without being spawned from an
existing project), and later as a TC member drafting early versions of
the Big Tent resolution.

When our community went through a period of intensive growth, as
companies joined and brought new features and formed new projects, I
helped organize the Oslo and release teams to cope with that growth by
implementing the liaison system and driving more use of automation.

More recently, as we have seen some leveling off, or even reduction,
of contributors, I worked with the documentation and requirements
teams to accommodate the contraction by shifting documentation from a
central location into team-owned repositories and by simplifying the
way we manage dependency settings in each project. These changes
reduced the friction for contributors to those projects.

Those experiences have prepared me to help us through this next
transition, as well.

My goals for this session are to have a smooth transition and to help
the TC work together as a team to achieve our shared objectives. I
have a few things I would like to change about the way we organize
ourselves based on patterns we have seen work well elsewhere,
including using sub-teams of TC members to manage our recurring
responsibilities more actively.

I look forward to talking with you about the future of the TC over the
next few weeks.

Change-Id: I35a4b1501b0019e79b34cc92f83af38ea4a41c26
Signed-off-by: Doug Hellmann <doug@doughellmann.com>