Clarify new project requirements for community engagement
From discussions on the mailing list[1] it seems that there is support
for the idea that projects that being with a code drop present a higher
risk of failing to attract interest outside of the initial developers.

This change aims to clearly communicate the TC's position to projects
considering applying to join OpenStack, by explicitly stating both that
code drops may be required to demonstrate traction in the community, and
that no such requirement for community engagement exists in general.

[1] http://lists.openstack.org/pipermail/openstack-dev/2018-April/129703.html

Change-Id: I8ee9bb08ee143402e2b8240e3cbe6ba5b0684596