move the team badge generation to the end of the build
Newer versions of Sphinx load the extensions before creating any of
the output directories, so the os.mkdir() call in the badge extension
was failing. Fix the problem by tying the badge step to the
build-finished event Sphinx emits when it reaches the end of its
build.

This change also uncaps Sphinx in the requirements file, since that
cap is not honored in the CI job anyway.

Change-Id: Ia6c232c5bc8c4015e00efe0b3b4182da2490acc7
Signed-off-by: Doug Hellmann <doug@doughellmann.com>