Castellan-compatible key store is a base service
Per the past year of discussions, culminating in the most recent
mailing list thread[*], it's apparent that providing a consistent
solution for storage of key material and similar secrets by security
features of various OpenStack services is in the best interests of
the project. By providing this guarantee in the base services set,
projects don't need to worry about implementing insecure fallback
alternatives or needlessly duplicating functionality to cope with
the lack of an already-available solution.

[*] http://lists.openstack.org/pipermail/openstack-dev/2018-May/130567.html

Change-Id: Ia46211f41726d5671bf28a632d17fc56965b6fcc