Include PTL E-mail address in rendered team pages
Display each current PTL's tracked E-mail address next to their name
and IRC nick. We've been tracking PTL contact E-mail addresses in
the governance repository for many cycles, but not presenting them
on the rendered
https://governance.openstack.org/tc/projects/<team>.html team detail
pages. The addresses themselves are far from secret and are already
"published" in the YAML version of our projects list, so there's
little to be gained by obscuring them from/in the HTML version.

Change-Id: I55d0ea058f645e8ba91cfc186eee00488ddbedcb
Story: #2001923
Task: #14450