Include TC E-mail addresses in roster
Track each current TC member's E-mail address next to their name and
IRC nick in the members roster and include it in the rendering on
the governance site, to make them easier for community members to
figure out how to contact in private when needed. We've included TC
candidate contact E-mail addresses in the elections repository for
many cycles, but not recorded them in governance once elected. The
addresses themselves are far from secret and are already "published"
in our election results records, so there's little to be gained by
omitting them from governance.

Change-Id: If7c22f056d33e874b606c48d86d1e4236eae8a99