add vice chair role to the tc charter
We currently make no provision for the chair to be able to delegate or
otherwise share their responsibilities. This change defines a new vice
chair role to give us a way to designate someone to act in a
supporting role for the active chair and to take over in the chair's
absence so that business does not need to halt when the chair is
unavailable.

Change-Id: I3e520eb0d900bc3d0552afbb5bcb9fd13fd0d09c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>