update link from python3-first goal to storyboard
We are using one story per team rather than a single story, so update
the link in the goal page to point to the query based on the tag
rather than linking to a single story.

Add a second link to the dashboard, which shows the stories in
different lanes based on whether they are completed or not.

Change-Id: I1588eaf7c00d855281f81c88f03366a4a3daf1b5
Signed-off-by: Doug Hellmann <doug@doughellmann.com>