Add openstack/os-ken to neutron
Neutron uses extensively the RYU framework (https://osrg.github.io/ryu/)
in its openflow related code. The group maintaining RYU has advised
the Neutron team that maintenance work will not continue after the
end of 2018. According to the discussion at Vancouver Summit,
the general consensus is to fork the RYU library under openstack namespace.
Neutron team will be responsible to maintain the fork.

To avoid confusion between the fork and the original library,
the neutron team picks a different name for the fork, which is "os-ken".

Depends-On: I1f94028a883eeba75d6697a4bcc4373a16569feb
Implements blueprint: ryu-framework-maintenace-transition

Change-Id: I4d320175dd1ab7dd66cbd8e5432c8f407bad75a5