Add a house rule about how to signal appointed PTLs
We've been dealing with how to make appointed PTLs official and
documented in a rather ad hoc fashion: something different each
cycle.

This change provides a house rule which puts the responsibility
and the documentation in the combined hands of the appointed PTL
and the TC that appointed them and not election handling (where
it doesn't make sense since no election happened).

Change-Id: Idf78e00c59d10528a2b4950f8baa93c6731973ea