Add a Vision statement for OpenStack clouds
Set out a vision for OpenStack clouds that clarifies, in long form, the
meaning of the mission statement at a technical level. Define what we
mean by 'cloud', describe how OpenStack must be different from other
cloud offerings, and list the design goals that we hope the OpenStack
system as a whole will achieve.

Change-Id: Ie988d87365cd2e190f2c6c13ee178f1590e4642d