Add Jay Faulkner as an extra-atc for ironic
Jay is a former ironic code contributor, but he never really
left ironic. To this day he can be found in the IRC channel
where he periodically raises through provoking discussion and
happilly tries to fill in context for the earlier days of ironic,
or help the occassional person seeking help.

Change-Id: I8ae1e7c0a35543b6f7ddb3bd1a4727dbcdf2d0b8