Move mistral-extra under mistral repos
This was released as its own thing in Newton, but since then has always
been included as the set of mistral releases. This updates the repo list
for the Mistral team to group it with the other main deliverables so it
matches what is actually being delivered. It also gets rid of a warning
in the release repo validation jobs that identify this mismatch.

Change-Id: I0103fa10e6f9c4d05c70fb914d2a123412f94ac0
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>