Add placement deliverable to nova
Add openstack/placement as repo under the nova project to indicate the
the home for placement as it is extracted from the nova code repository
to its own.

The new repo [1] was created by the (merged) change at the depends-on
below, which seeded the repo from a filtered clone of nova. See that
change for more details on the disposition of the repo and the immediate
plans.

[1] https://git.openstack.org/cgit/openstack/placement

Change-Id: I227f7d109de798fea65569e8a93de8b233be591b
Depends-On: https://review.openstack.org/598362