Add a note about deployment projects for upgrade-checkers goal
Deployment projects are not the primary focus nor a goal
for integration of the upgrade-checkers that get added
to service projects in Stein. This change adds a note about
deployment projects for the upgrade-checkers goal that while
those projects are not a target for the goal, integration
with them is obviously a good thing.

Change-Id: I2ba644d5d89331754bf2ff25e556b92b72024ee9